// src/app/interfaces/common/page-faq.interface.ts
export interface PageFaq {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  informations?: any[];
  description?: string;
  title?: string;
  descriptionEn?: string;
  titleEn?: string;
  pages?: string;
  subTitle?: string;
  shortDescription?: string;
  subTitleEn?: string;
  shortDescriptionEn?: string;
  slug?: string;
  createdAt?: Date;
  updatedAt?: Date;
}